Code maintenance, no functional changes
