+++ GitUI/UserControls/ExListView.cs
@@ -12,60 +12,14 @@ using JetBrains.Annotations;
 
 namespace GitUI.UserControls
 {
-    public enum ListViewGroupState : uint
-    {
-        /// <summary>
-        /// Groups are expanded, the group name is displayed,
-        /// and all items in the group are displayed.
-        /// </summary>
-        Normal = 0,
-
-        /// <summary>
-        /// The group is collapsed.
-        /// </summary>
-        Collapsed = 1,
-
-        /// <summary>
-        /// The group is hidden.
-        /// </summary>
-        Hidden = 2,
-
-        /// <summary>
-        /// Version 6.00 and Windows Vista. The group does not display a header.
-        /// </summary>
-        NoHeader = 4,
-
-        /// <summary>
-        /// Version 6.00 and Windows Vista. The group can be collapsed.
-        /// </summary>
-        Collapsible = 8,
-
-        /// <summary>
-        /// Version 6.00 and Windows Vista. The group has keyboard focus.
-        /// </summary>
-        Focused = 16,
-
-        /// <summary>
-        /// Version 6.00 and Windows Vista. The group is selected.
-        /// </summary>
-        Selected = 32,
-
-        /// <summary>
-        /// Version 6.00 and Windows Vista. The group displays only a portion of its items.
-        /// </summary>
-        SubSeted = 64,
-
-        /// <summary>
-        /// Version 6.00 and Windows Vista. The subset link of the group has keyboard focus.
-        /// </summary>
-        SubSetLinkFocused = 128
-    }
-
     internal class ExListView : NativeListView
     {
         private static readonly PropertyInfo ListViewGroupIdProperty;
         private static readonly PropertyInfo ListViewGroupListProperty;
         private static readonly PropertyInfo ListViewDefaultGroupProperty;
+
+        /// <summary> Position in <see cref="ListView.Groups"/> collection after the technical "Default" group </summary>
+        private int _minGroupInsertionIndex;
 
         /// <summary>
         /// Occurs when the user presses mouse button in a <see cref="ListViewGroup"/> within the list view control.
@@ -97,52 +51,43 @@ namespace GitUI.UserControls
             IsGroupStateSupported = EnvUtils.RunningOnWindows() && Environment.OSVersion.Version.Major >= 6;
         }
 
-        /// <summary>
-        /// Call this method once after <see cref="ListView.Groups"/> collection was cleared (or created)
-        /// before making any calls to <see cref="ListViewGroupCollection.Insert"/>
-        ///
-        /// <see cref="ListViewGroupCollection.Add(System.Windows.Forms.ListViewGroup)"/> calls must
-        /// not be used after calling this method
-        /// </summary>
-        private void BeginGroupInsertion()
-        {
-            // .NET ListView.Groups.Insert(...) implementation has a bug
-            // It does not count the technical "Default" group from ListView when passing inserted
-            // group index to native Win32 ListView.
-
-            // ListView.Groups.Add(...) implementation on the other hand does count the
-            // technical "Default" group correctly therefore it becomes broken after calling this method
-
-            var defaultGroup = ListViewDefaultGroupProperty.GetValue(this);
-            var list = (ArrayList)ListViewGroupListProperty.GetValue(Groups);
-            if (list.Count > 0 && list[0] == defaultGroup)
-            {
-                throw new InvalidOperationException($"{nameof(BeginGroupInsertion)} was already called");
-            }
-
-            list.Insert(0, defaultGroup);
-            _minGroupInsertionIndex = 1;
-        }
-
-        private void EndGroupInsertion()
-        {
-            var defaultGroup = ListViewDefaultGroupProperty.GetValue(this);
-            var list = (ArrayList)ListViewGroupListProperty.GetValue(Groups);
-
-            if (list.Count == 0 || list[0] != defaultGroup)
-            {
-                throw new InvalidOperationException($"{nameof(BeginGroupInsertion)} was not called before");
-            }
-
-            list.RemoveAt(0);
-            _minGroupInsertionIndex = 0;
-        }
-
         [Category("Behavior"), DefaultValue(false)]
         public bool AllowCollapseGroups { get; set; }
 
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
         private bool IsGroupStateSupported { get; }
+
+        [CanBeNull]
+        public ListViewGroupHitInfo GetGroupHitInfo(Point location)
+        {
+            var info = new NativeMethods.LVHITTESTINFO
+            {
+                pt = location
+            };
+
+            var handleRef = new HandleRef(this, Handle);
+            if (NativeMethods.SendMessage(handleRef, NativeMethods.LVM_SUBITEMHITTEST, (IntPtr)(-1), ref info) == new IntPtr(-1))
+            {
+                return null;
+            }
+
+            if ((info.flags & NativeMethods.LVHITTESTFLAGS.LVHT_EX_GROUP_HEADER) == 0)
+            {
+                return null;
+            }
+
+            foreach (ListViewGroup group in Groups)
+            {
+                var groupId = GetGroupId(group);
+                if (info.iItem == groupId)
+                {
+                    bool isCollapseButton = (info.flags & NativeMethods.LVHITTESTFLAGS.LVHT_EX_GROUP_COLLAPSE) > 0;
+                    return new ListViewGroupHitInfo(group, isCollapseButton, location);
+                }
+            }
+
+            return null;
+        }
 
         /// <summary>
         /// Modifies <see cref="ListView.Groups"/> collection to match supplied one.
@@ -184,157 +129,6 @@ namespace GitUI.UserControls
 
             IEnumerable<ListViewGroup> GetGroups() =>
                 Groups.Cast<ListViewGroup>().Skip(_minGroupInsertionIndex);
-        }
-
-        #region Win32 Apis
-
-        private static class NativeMethods
-        {
-            [DllImport("user32", CharSet = CharSet.Auto)]
-            public static extern IntPtr SendMessage(HandleRef hWnd,
-                                                   int msg,
-                                                   IntPtr wParam,
-                                                   ref LVHITTESTINFO lParam);
-
-            [DllImport("user32", CharSet = CharSet.Auto)]
-            public static extern IntPtr SendMessage(HandleRef hWnd,
-                                                   int msg,
-                                                   IntPtr wParam,
-                                                   ref LVGROUP lParam);
-
-            [DllImport("user32.dll", CharSet = CharSet.Auto)]
-            public static extern int GetScrollPos(IntPtr hWnd, int nBar);
-
-            #region Windows constants
-
-            public const int WM_LBUTTONDOWN = 0x0201;
-            public const int WM_LBUTTONUP = 0x0202;
-            public const int WM_RBUTTONDOWN = 0x0204;
-            public const int WM_RBUTTONUP = 0x0205;
-
-            public const int WM_VSCROLL = 0x115;
-            public const int WM_MOUSEWHEEL = 0x020A;
-            public const int WM_KEYDOWN = 0x0100;
-            public const int SB_VERT = 1;
-
-            public const int LVM_FIRST = 0x1000;
-            public const int LVM_HITTEST = LVM_FIRST + 18;
-            public const int LVM_SETGROUPINFO = LVM_FIRST + 147;
-            public const int LVM_SUBITEMHITTEST = LVM_FIRST + 57;
-            public const int LVM_INSERTGROUP = LVM_FIRST + 145;
-
-            #endregion
-
-            [StructLayout(LayoutKind.Sequential)]
-            public readonly struct POINT
-            {
-                public readonly int X;
-                public readonly int Y;
-
-                public POINT(int x, int y)
-                {
-                    X = x;
-                    Y = y;
-                }
-
-                public static implicit operator System.Drawing.Point(POINT p) => new System.Drawing.Point(p.X, p.Y);
-                public static implicit operator POINT(System.Drawing.Point p) => new POINT(p.X, p.Y);
-            }
-
-            [StructLayout(LayoutKind.Sequential)]
-            public struct NMHDR
-            {
-                public IntPtr hwndFrom;
-                public IntPtr idFrom;
-                public int code;
-            }
-
-            /// <summary>
-            /// see http://msdn.microsoft.com/en-us/library/bb774754%28v=VS.85%29.aspx
-            /// </summary>
-            [StructLayout(LayoutKind.Sequential)]
-            public struct LVHITTESTINFO
-            {
-                public POINT pt;
-                public LVHITTESTFLAGS flags;
-                public int iItem;
-                public int iSubItem;
-
-                // Vista/Win7+
-                public int iGroup;
-            }
-
-            /// <summary>
-            /// see http://msdn.microsoft.com/en-us/library/bb774754%28v=VS.85%29.aspx
-            /// </summary>
-            [Flags]
-            public enum LVHITTESTFLAGS : uint
-            {
-                LVHT_NOWHERE = 0x00000001,
-                LVHT_ONITEMICON = 0x00000002,
-                LVHT_ONITEMLABEL = 0x00000004,
-                LVHT_ONITEMSTATEICON = 0x00000008,
-                LVHT_ONITEM = LVHT_ONITEMICON | LVHT_ONITEMLABEL | LVHT_ONITEMSTATEICON,
-                LVHT_ABOVE = 0x00000008,
-                LVHT_BELOW = 0x00000010,
-                LVHT_TORIGHT = 0x00000020,
-                LVHT_TOLEFT = 0x00000040,
-
-                // Vista/Win7+ only
-                LVHT_EX_GROUP_HEADER = 0x10000000,
-                LVHT_EX_GROUP_FOOTER = 0x20000000,
-                LVHT_EX_GROUP_COLLAPSE = 0x40000000,
-                LVHT_EX_GROUP_BACKGROUND = 0x80000000,
-                LVHT_EX_GROUP_STATEICON = 0x01000000,
-                LVHT_EX_GROUP_SUBSETLINK = 0x02000000,
-            }
-
-            public enum ListViewGroupMask : uint
-            {
-                None = 0x00000,
-                Header = 0x00001,
-                Footer = 0x00002,
-                State = 0x00004,
-                Align = 0x00008,
-                GroupId = 0x00010,
-                SubTitle = 0x00100,
-                Task = 0x00200,
-                DescriptionTop = 0x00400,
-                DescriptionBottom = 0x00800,
-                TitleImage = 0x01000,
-                ExtendedImage = 0x02000,
-                Items = 0x04000,
-                Subset = 0x08000,
-                SubsetItems = 0x10000
-            }
-
-            [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-            public struct LVGROUP
-            {
-                public int CbSize;
-                public ListViewGroupMask Mask;
-                [MarshalAs(UnmanagedType.LPWStr)]
-                public string PszHeader;
-                public int CchHeader;
-                [MarshalAs(UnmanagedType.LPWStr)]
-                public string PszFooter;
-                public int CchFooter;
-                public int IGroupId;
-                public int StateMask;
-                public ListViewGroupState State;
-                public uint UAlign;
-            }
-        }
-        #endregion
-
-        protected override void SetBoundsCore(int x, int y, int width, int height, BoundsSpecified specified)
-        {
-            // Prevent flickering horizontal scrollbar when control width is reduced,
-            // by preventing redrawing the control before column width is adjusted to new width in
-            // some event handler e.g. ClientSizeChanged.
-            BeginUpdate();
-            base.SetBoundsCore(x, y, width, height, specified);
-            EndUpdate();
         }
 
         protected override void WndProc(ref Message m)
@@ -495,36 +289,55 @@ namespace GitUI.UserControls
             }
         }
 
-        [CanBeNull]
-        public ListViewGroupHitInfo GetGroupHitInfo(Point location)
+        /// <summary>
+        /// Call this method once after <see cref="ListView.Groups"/> collection was cleared (or created)
+        /// before making any calls to <see cref="ListViewGroupCollection.Insert"/>
+        ///
+        /// <see cref="ListViewGroupCollection.Add(System.Windows.Forms.ListViewGroup)"/> calls must
+        /// not be used after calling this method
+        /// </summary>
+        private void BeginGroupInsertion()
         {
-            var info = new NativeMethods.LVHITTESTINFO
-            {
-                pt = location
-            };
+            // .NET ListView.Groups.Insert(...) implementation has a bug
+            // It does not count the technical "Default" group from ListView when passing inserted
+            // group index to native Win32 ListView.
 
-            var handleRef = new HandleRef(this, Handle);
-            if (NativeMethods.SendMessage(handleRef, NativeMethods.LVM_SUBITEMHITTEST, (IntPtr)(-1), ref info) == new IntPtr(-1))
+            // ListView.Groups.Add(...) implementation on the other hand does count the
+            // technical "Default" group correctly therefore it becomes broken after calling this method
+
+            var defaultGroup = ListViewDefaultGroupProperty.GetValue(this);
+            var list = (ArrayList)ListViewGroupListProperty.GetValue(Groups);
+            if (list.Count > 0 && list[0] == defaultGroup)
             {
-                return null;
+                throw new InvalidOperationException($"{nameof(BeginGroupInsertion)} was already called");
             }
 
-            if ((info.flags & NativeMethods.LVHITTESTFLAGS.LVHT_EX_GROUP_HEADER) == 0)
+            list.Insert(0, defaultGroup);
+            _minGroupInsertionIndex = 1;
+        }
+
+        private void EndGroupInsertion()
+        {
+            var defaultGroup = ListViewDefaultGroupProperty.GetValue(this);
+            var list = (ArrayList)ListViewGroupListProperty.GetValue(Groups);
+
+            if (list.Count == 0 || list[0] != defaultGroup)
             {
-                return null;
+                throw new InvalidOperationException($"{nameof(BeginGroupInsertion)} was not called before");
             }
 
-            foreach (ListViewGroup group in Groups)
-            {
-                var groupId = GetGroupId(group);
-                if (info.iItem == groupId)
-                {
-                    bool isCollapseButton = (info.flags & NativeMethods.LVHITTESTFLAGS.LVHT_EX_GROUP_COLLAPSE) > 0;
-                    return new ListViewGroupHitInfo(group, isCollapseButton, location);
-                }
-            }
+            list.RemoveAt(0);
+            _minGroupInsertionIndex = 0;
+        }
 
-            return null;
+        protected override void SetBoundsCore(int x, int y, int width, int height, BoundsSpecified specified)
+        {
+            // Prevent flickering horizontal scrollbar when control width is reduced,
+            // by preventing redrawing the control before column width is adjusted to new width in
+            // some event handler e.g. ClientSizeChanged.
+            BeginUpdate();
+            base.SetBoundsCore(x, y, width, height, specified);
+            EndUpdate();
         }
 
         private static int GetGroupId(ListViewGroup listViewGroup)
@@ -544,7 +357,142 @@ namespace GitUI.UserControls
             return -1;
         }
 
-        /// <summary> Position in <see cref="ListView.Groups"/> collection after the technical "Default" group </summary>
-        private int _minGroupInsertionIndex;
+        private static class NativeMethods
+        {
+            [DllImport("user32", CharSet = CharSet.Auto)]
+            public static extern IntPtr SendMessage(HandleRef hWnd,
+                                                   int msg,
+                                                   IntPtr wParam,
+                                                   ref LVHITTESTINFO lParam);
+
+            [DllImport("user32", CharSet = CharSet.Auto)]
+            public static extern IntPtr SendMessage(HandleRef hWnd,
+                                                   int msg,
+                                                   IntPtr wParam,
+                                                   ref LVGROUP lParam);
+
+            [DllImport("user32.dll", CharSet = CharSet.Auto)]
+            public static extern int GetScrollPos(IntPtr hWnd, int nBar);
+
+            #region Windows constants
+
+            public const int WM_LBUTTONDOWN = 0x0201;
+            public const int WM_LBUTTONUP = 0x0202;
+            public const int WM_RBUTTONDOWN = 0x0204;
+            public const int WM_RBUTTONUP = 0x0205;
+
+            public const int WM_VSCROLL = 0x115;
+            public const int WM_MOUSEWHEEL = 0x020A;
+            public const int WM_KEYDOWN = 0x0100;
+            public const int SB_VERT = 1;
+
+            public const int LVM_FIRST = 0x1000;
+            public const int LVM_HITTEST = LVM_FIRST + 18;
+            public const int LVM_SETGROUPINFO = LVM_FIRST + 147;
+            public const int LVM_SUBITEMHITTEST = LVM_FIRST + 57;
+            public const int LVM_INSERTGROUP = LVM_FIRST + 145;
+
+            #endregion
+
+            [StructLayout(LayoutKind.Sequential)]
+            public readonly struct POINT
+            {
+                public readonly int X;
+                public readonly int Y;
+
+                public POINT(int x, int y)
+                {
+                    X = x;
+                    Y = y;
+                }
+
+                public static implicit operator System.Drawing.Point(POINT p) => new System.Drawing.Point(p.X, p.Y);
+                public static implicit operator POINT(System.Drawing.Point p) => new POINT(p.X, p.Y);
+            }
+
+            [StructLayout(LayoutKind.Sequential)]
+            public struct NMHDR
+            {
+                public IntPtr hwndFrom;
+                public IntPtr idFrom;
+                public int code;
+            }
+
+            /// <summary>
+            /// see http://msdn.microsoft.com/en-us/library/bb774754%28v=VS.85%29.aspx
+            /// </summary>
+            [StructLayout(LayoutKind.Sequential)]
+            public struct LVHITTESTINFO
+            {
+                public POINT pt;
+                public LVHITTESTFLAGS flags;
+                public int iItem;
+                public int iSubItem;
+
+                // Vista/Win7+
+                public int iGroup;
+            }
+
+            /// <summary>
+            /// see http://msdn.microsoft.com/en-us/library/bb774754%28v=VS.85%29.aspx
+            /// </summary>
+            [Flags]
+            public enum LVHITTESTFLAGS : uint
+            {
+                LVHT_NOWHERE = 0x00000001,
+                LVHT_ONITEMICON = 0x00000002,
+                LVHT_ONITEMLABEL = 0x00000004,
+                LVHT_ONITEMSTATEICON = 0x00000008,
+                LVHT_ONITEM = LVHT_ONITEMICON | LVHT_ONITEMLABEL | LVHT_ONITEMSTATEICON,
+                LVHT_ABOVE = 0x00000008,
+                LVHT_BELOW = 0x00000010,
+                LVHT_TORIGHT = 0x00000020,
+                LVHT_TOLEFT = 0x00000040,
+
+                // Vista/Win7+ only
+                LVHT_EX_GROUP_HEADER = 0x10000000,
+                LVHT_EX_GROUP_FOOTER = 0x20000000,
+                LVHT_EX_GROUP_COLLAPSE = 0x40000000,
+                LVHT_EX_GROUP_BACKGROUND = 0x80000000,
+                LVHT_EX_GROUP_STATEICON = 0x01000000,
+                LVHT_EX_GROUP_SUBSETLINK = 0x02000000,
+            }
+
+            public enum ListViewGroupMask : uint
+            {
+                None = 0x00000,
+                Header = 0x00001,
+                Footer = 0x00002,
+                State = 0x00004,
+                Align = 0x00008,
+                GroupId = 0x00010,
+                SubTitle = 0x00100,
+                Task = 0x00200,
+                DescriptionTop = 0x00400,
+                DescriptionBottom = 0x00800,
+                TitleImage = 0x01000,
+                ExtendedImage = 0x02000,
+                Items = 0x04000,
+                Subset = 0x08000,
+                SubsetItems = 0x10000
+            }
+
+            [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+            public struct LVGROUP
+            {
+                public int CbSize;
+                public ListViewGroupMask Mask;
+                [MarshalAs(UnmanagedType.LPWStr)]
+                public string PszHeader;
+                public int CchHeader;
+                [MarshalAs(UnmanagedType.LPWStr)]
+                public string PszFooter;
+                public int CchFooter;
+                public int IGroupId;
+                public int StateMask;
+                public ListViewGroupState State;
+                public uint UAlign;
+            }
+        }
     }
 }
